Only send TerminationAck to the leader that requested it.
Previously, when issuing a shutdown, via the control toggle, on the
leader of a cluster and then immediately closing its ConsensusModule, it
was possible that a new leader was elected and a follower node (in both
terms) could send a TerminationAck requested by the first leader to the
second leader. As the second leader was not expecting to receive this
TerminationAck, it would throw a NPE when attempting to access
`ConsensusModuleAgent#clusterTermination`.

Now, the follower stores the `leadershipTermId` of the leader that has
requested a termination and will only send the TerminationAck if an
election has not occurred in the interim.

I have included a test which fails ~ 50% of the time before the
changes to `ConsensusModuleAgent` are applied. It was necessary to take
fine-grained control over the progress of certain threads to recreate
the problem. Therefore, I introduced a helper class `MethodCallBlocker`,
which supports blocking a particular thread when it enters a specific
method.