Disable test due to intermittency on mac.
This test helped to prove that a fix worked to avoid an NPE due a
shutdown followed by immediate leader close.

There was some discussion here: https://weareadaptive.slack.com/archives/C07LRBUFY83/p1746783266421059
(Sorry, it is only visible to Adaptive employees)

It may rot, as it references methods by strings for the purposes of
blocking the progress of certain threads. Therefore, we may want to
consider removing it.