Set `terminationLeadershipTermId` whenever `terminationPosition` is set.
Currently, `terminationLeadershipTermId` is only set when the follower
receives a `terminationPosition` from the leader. It uses it to
determine whether the "terminating leader" is still the leader when
ready to terminate and deciding whether to send a termination ack.

We thought it was best that the field always be set whenever
`terminationPosition` is set, i.e., even on the leader, when handling
control signals, as it will hopefully be less surprising / easier to
rely on.