aeron_atomic64_gcc_x86_64.h acquire/release fixes
On the X86, every store has release semantics and every load
has acquire semantics.

The only thing that needs to be controlled is the compiler and
that can be done using the

__asm__ volatile("" ::: "memory");

This is inline with Linux:

https://github.com/torvalds/linux/blob/master/arch/x86/include/asm/barrier.h

And the barrier is defined here:

https://github.com/torvalds/linux/blob/master/include/linux/compiler.h

/* The "volatile" is due to gcc bugs */

Perhaps a good idea if we add the __volatile__ as well.

So there is no need for a fake volatile load or store in combination
with a compiler barrier. It is sufficient to only have the compiler
barrier.