scan_back_to_confirm_still_zeroed underflow problem
It can happen that `i` can underflow and wraps around to a huge value
because

i -= AERON_RB_ALIGNMENT

E.g. if `i` is 66 and AERON_RB_ALIGNMENT, then after the first round
`i` will be 2. After the second round `i` will be `SIZE_MAX-62` which
is a huge number.

So at that point you run into UB.

Currently is fixed by adding a guard before the decrement of `i`.